Automatically backed up by Learn

+++ db/migrate/20190406152622_create_users.rb
@@ -3,7 +3,7 @@ class CreateUsers < ActiveRecord::Migration[5.1]
     create_table :users do |element|
       element.string :username
       element.string :password
-      element.float :balance
+      element.decimal :balance
     end
   end
 end
